Move StatusEvents to run_models/events
